Remove a long-ago stupid hack that was screwing my happiness.  The details are these:  When I added the ability to transmit attribute lists along with CM messages I had to add an attribute length element to the message header. The header already consisted of a 4-byte magic number and a 4-byte data length value.  I needed to preserve the situation where the actual data was aligned on an 8-byte boundary.  I didn't want to try to squeeze the attr length into the 8 header bytes I was already using, but going to 16 bytes seemed like overkill.  So I committed an ugly hack.  I took advantage of the fact that the CMDataAvailable routine looked at the initial 4 bytes first (reading them separately on the cmsockets transport), and only later read the rest of the message package.  My hack was that I transmitted a 12-byte header for CM-message-with-attributes, but on the receiving side I examined the magic number (first 4 bytes) and then manipulated the receive buffer to add 4-bytes of "padding" that were necessary to preserve alignment, but weren't transmitted.  Ugly, but workable.
Next, came the UDP transport where the whole message came in as a single
packet.  I had a problem in that the 12-byte header here made the data
mis-aligned if I read it in on an 8-byte boundary.  Fortunately the 
MSG_PEEK function let me look at the first 4 bytes of the message, check the
magic number and provide a suitably-misaligned buffer so that the 12-byte
header made the data perfectly aligned anyway.  This was an uglier hack, but
still workable.

Now, with the IB transport being developed, the wheels have come off this
particular Mad Max mobile.  I've changed to a proper 16-byte header for this
situation, removing the hack from the UDP transport and the ugly buffer
manipulation that happened as well.  Now all incoming messages can come in
on an 8-byte boundary and be processed with their natural alignment.

NOTE:  This is effectively a change in the wire format that we use for
CM-level message with attributes.  Old and new libraries will not properly
communicate.  However, I think we can risk introducing the change without
trying to ease it.  Few things use CM-level messages with attributes (as
we're mostly EVPath-based and EVpath-level messages are unaffected.
Additionally, as long as you're using a DLL on CoC based machines, both
sides will be updated.  The only real issue is CM-level messages with
attributes where one side has a new DLL and the other has not.  Maybe we'll
see that somewhere, but I've got my fingers crossed that we won't.


git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@8825 8ba6a8b9-845b-0410-94f0-974be2f5df47