RDMA write now works ALMOST correctly. However there is no way for the receiving side to know when a RDMA_WRITE operation has completed (seriously ofed? how stupid are you guys? fucking retards).
I need to think a bit about the solution. The easiest way would be to wait
for a socket message from the sender in data_available. Since the sending
side knows exactly when the RDMA_WRITE has completed correctly this will be
correct but will require an additional message and data available will block
for the entirety of the transfer. That can't be good!

The better way I think would be to let data available return without
data. Then periodically (using the periodic task mechanism) we can check for
a message on the dataqp. This message will be sent over IB (not ipoib) once
the write is completed. By letting the notification happen asynchronously
we'll be able to return from data available really fast and we can have
multiple incoming requests pending at the same time.


git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@8951 8ba6a8b9-845b-0410-94f0-974be2f5df47