Sending over RDMA is now working correctly.
1) Fixed the memory corruption that was happening when we
free the remote host string. 

2) Changed the name of the internal functions from socket_ to ib_

3) Added a waitoncq function to simplify waiting for notification. 

4) Removed SRQs since we only talk to one guy on a RC qp.

5) Added some more debug statements. These need to removed ofcourse.

Still on the list: Fix the CMtake_buffer issue. Add more buffer management
to the system. Optimize how we allocate and manage the buffer in
data_available. On the sending side we don't have a clean solution for
handling a lot of iovecs. As long as the iovec count is less than 62 we are
good. If we increase beyond that we'll have to split it into multiple work. 

Possible things to do: switch to RDMA_CM for connection management. Test for
multiple sending nodes.


git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@8966 8ba6a8b9-845b-0410-94f0-974be2f5df47