Add a CM-level handshake to connection initiation.  At this time, the handshake just contains the (randomly generated) identifier for the format server in use by each side.  This update also rewrites the format preloading code so that a CM will preload (send down server-style format information) formats IFF 1) it is self-hosting formats (not communicating with an external format server), 2) the remote CM is self-hosting formats, or 3) the two CMs are communicating with different format servers.  The result of this change is that we can freely communicate between self-hosted, and differently hosted CMs.
Soon to be added will be a random ID for each communicating CM.  The idea
is that this latter addition might help us with things like identifying a
reliable transport down which to preload formats if we are working with
unreliable transports.  This latter idea may not be realistic, as we'd have
to contrive a mechanism through which we could make sure the
reliably-transmitted format was read and handled before the possibly
unreliable message, but it's not impossible.  Still wouldn't help us with
multicast transports, etc.  But tossing in a few extra bytes in the
handshake is cheap at this point.


git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@11559 8ba6a8b9-845b-0410-94f0-974be2f5df47