Hopefully take care of a little issue that Jai was running into.  In order to keep track of what event_item was associated with the currently-running handler (so that we could find it in EVtake_event_buffer()), we were storing it in cm->evp->current_event_item.  The trouble with this is that there is *not*, necessarily, just one currently executing handler.  In fact, handlers could be being executed by any thread that does an EVsubmit() and so could be concurrent (at least at the point at which locks are dropped, during actual handler execution).   Additionally, if one does an EVsubmit() inside a handler, this could itself recurse as we try to process that event before returning from the EVsubmit, with the result that there could be several handlers on the stack of a single process.
So, the idea of a single evp->current_event_item is bad.  Stupid bad.  It is
replaced with a list.  That list is manipulated while we are holding the
locks, so concurrency is not at issue.  However, we must search the list in
EVtake_event_buffer() to find out which item matches the buffer address we
are given, *and* we must search the list when we return from a handler
because while our event might have been first in the list when we called the
handler, the state of the list will perhaps have changed before that handler
returns.


git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@15118 8ba6a8b9-845b-0410-94f0-974be2f5df47