Provide EVdfg_dump_graph(EVdfg_state_type which, EVdfg dfg), a mechanism to dump a graph ML representation of the DFG.
As a means of discerning whether the programmer has created the stone graph
that is intended, this call provides a mechanism to print (to stdout) a
GraphML representation of the stones and links as graph vertices and
directed edges.  At present, the output is best viewed in the yEd graph
editor.  Copy the output from the screen, paste to a file, open in yEd and
choose a layout scheme (try auto-layout).

Currently we set a node label (specific to yEd) that is just S%d based on
the stone id.  Links also have the appropriate output port property set,
though this does not affect their graphical display.


git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@23634 8ba6a8b9-845b-0410-94f0-974be2f5df47